+++ pkg/reqctx/info.go
@@ -1,0 +1,7 @@
+package reqctx
+
+type RequestProcessInfo struct {
+	Command       string
+	Params        string
+	ContentLength int
+}